+++ lib/providers/active_wallets.dart
@@ -97,7 +97,7 @@ class ActiveWallets with ChangeNotifier {
   Future<String?> getAddressFromDerivationPath(
       String identifier, int account, int chain, int address,
       [master = false]) async {
-    final network = AvailableCoins().getSpecificCoin(identifier).networkType;
+    final network = AvailableCoins.getSpecificCoin(identifier).networkType;
     var hdWallet = HDWallet.fromSeed(
       seedPhraseUint8List(await seedPhrase),
       network: network,
@@ -135,7 +135,7 @@ class ActiveWallets with ChangeNotifier {
 
   Future<void> generateUnusedAddress(String identifier) async {
     var openWallet = getSpecificCoinWallet(identifier);
-    final network = AvailableCoins().getSpecificCoin(identifier).networkType;
+    final network = AvailableCoins.getSpecificCoin(identifier).networkType;
     var hdWallet = HDWallet.fromSeed(
       seedPhraseUint8List(await seedPhrase),
       network: network,
@@ -313,6 +313,10 @@ class ActiveWallets with ChangeNotifier {
       //check if that tx is already in the db
       var txInWallet = openWallet.transactions;
       var isInWallet = false;
+      final decimalProduct = AvailableCoins.getDecimalProduct(
+        identifier: identifier,
+      );
+
       for (var walletTx in txInWallet) {
         if (walletTx.txid == tx['txid']) {
           isInWallet = true;
@@ -341,23 +345,24 @@ class ActiveWallets with ChangeNotifier {
           for (var vOut in voutList) {
             final asMap = vOut as Map;
             if (asMap['scriptPubKey']['type'] != 'nulldata') {
-              asMap['scriptPubKey']['addresses'].forEach((addr) {
-                if (openWallet.addresses.firstWhereOrNull(
-                        (element) => element.address == addr) !=
-                    null) {
-                  //address is ours, add new tx
-                  final txValue = (vOut['value'] * 1000000).toInt();
+              asMap['scriptPubKey']['addresses'].forEach(
+                (addr) {
+                  if (openWallet.addresses.firstWhereOrNull(
+                          (element) => element.address == addr) !=
+                      null) {
+                    //address is ours, add new tx
+                    final txValue = (vOut['value'] * decimalProduct).toInt();
 
-                  //increase notification value for addr
-                  final addrInWallet = openWallet.addresses
-                      .firstWhere((element) => element.address == addr);
-                  addrInWallet.newNotificationBackendCount =
-                      addrInWallet.notificationBackendCount + 1;
-                  openWallet.save();
+                    //increase notification value for addr
+                    final addrInWallet = openWallet.addresses
+                        .firstWhere((element) => element.address == addr);
+                    addrInWallet.newNotificationBackendCount =
+                        addrInWallet.notificationBackendCount + 1;
+                    openWallet.save();
 
-                  //write tx
-                  openWallet.putTransaction(
-                    WalletTransaction(
+                    //write tx
+                    openWallet.putTransaction(
+                      WalletTransaction(
                         txid: tx['txid'],
                         timestamp: tx['blocktime'] ?? 0,
                         value: txValue,
@@ -367,10 +372,12 @@ class ActiveWallets with ChangeNotifier {
                         broadCasted: true,
                         confirmations: tx['confirmations'] ?? 0,
                         broadcastHex: '',
-                        opReturn: ''),
-                  );
-                }
-              });
+                        opReturn: '',
+                      ),
+                    );
+                  }
+                },
+              );
             }
           }
 
@@ -530,7 +537,7 @@ class ActiveWallets with ChangeNotifier {
     String identifier,
     String address,
   ) async {
-    var network = AvailableCoins().getSpecificCoin(identifier).networkType;
+    var network = AvailableCoins.getSpecificCoin(identifier).networkType;
     var openWallet = getSpecificCoinWallet(identifier);
     var walletAddress = openWallet.addresses
         .firstWhereOrNull((element) => element.address == address);
@@ -562,7 +569,7 @@ class ActiveWallets with ChangeNotifier {
   Future<Map> buildTransaction({
     required String identifier,
     required String address,
-    required String amount,
+    required double amount,
     required int fee,
     String opReturn = '',
     bool firstPass = true,
@@ -570,8 +577,12 @@ class ActiveWallets with ChangeNotifier {
     String paperWalletPrivkey = '',
   }) async {
     //convert amount
-    var _txAmount = (double.parse(amount) * 1000000).toInt();
-    var openWallet = getSpecificCoinWallet(identifier);
+    final _decimalProduct = AvailableCoins.getDecimalProduct(
+      identifier: identifier,
+    );
+
+    var _txAmount = (amount * _decimalProduct).toInt();
+    var _openWallet = getSpecificCoinWallet(identifier);
     var _hex = '';
     var _destroyedChange = 0;
 
@@ -589,7 +600,7 @@ class ActiveWallets with ChangeNotifier {
 
     //check if tx needs change
     var _needsChange = true;
-    if (_txAmount == openWallet.balance || paperWalletUtxos != null) {
+    if (_txAmount == _openWallet.balance || paperWalletUtxos != null) {
       _needsChange = false;
       LoggerWrapper.logInfo(
         'ActiveWallets',
@@ -604,14 +615,14 @@ class ActiveWallets with ChangeNotifier {
     }
 
     //define utxo pool
-    var utxoPool = paperWalletUtxos ?? openWallet.utxos;
+    var utxoPool = paperWalletUtxos ?? _openWallet.utxos;
 
-    if (_txAmount <= openWallet.balance || paperWalletUtxos != null) {
+    if (_txAmount <= _openWallet.balance || paperWalletUtxos != null) {
       if (utxoPool.isNotEmpty) {
         //find eligible input utxos
         var _totalInputValue = 0;
         var inputTx = <WalletUtxo>[];
-        var coin = AvailableCoins().getSpecificCoin(identifier);
+        var coin = AvailableCoins.getSpecificCoin(identifier);
 
         for (var utxo in utxoPool) {
           if (utxo.value > 0) {
@@ -627,7 +638,7 @@ class ActiveWallets with ChangeNotifier {
           }
         }
 
-        var coinParams = AvailableCoins().getSpecificCoin(identifier);
+        var coinParams = AvailableCoins.getSpecificCoin(identifier);
         var network = coinParams.networkType;
 
         //start building tx
@@ -674,7 +685,7 @@ class ActiveWallets with ChangeNotifier {
           for (var inputUtxo in inputTx) {
             var inputKey = inputTx.indexOf(inputUtxo);
             //find key to that utxo
-            for (var walletAddr in openWallet.addresses) {
+            for (var walletAddr in _openWallet.addresses) {
               if (walletAddr.address == inputUtxo.address) {
                 var wif = paperWalletUtxos != null
                     ? paperWalletPrivkey
@@ -706,7 +717,7 @@ class ActiveWallets with ChangeNotifier {
         final intermediate = tx.build();
         var number = ((intermediate.txSize) / 1000 * coin.feePerKb)
             .toStringAsFixed(coin.fractions);
-        var asDouble = double.parse(number) * 1000000;
+        var asDouble = double.parse(number) * _decimalProduct;
         var requiredFeeInSatoshis = asDouble.toInt();
 
         LoggerWrapper.logInfo(
@@ -782,7 +793,7 @@ class ActiveWallets with ChangeNotifier {
   }
 
   String getScriptHash(String identifier, String address) {
-    var network = AvailableCoins().getSpecificCoin(identifier).networkType;
+    var network = AvailableCoins.getSpecificCoin(identifier).networkType;
     var script = Address.addressToOutputScript(address, network);
     var hash = sha256.convert(script).toString();
     return (reverseString(hash));