+++ lib/screens/server_settings/server_settings.dart
@@ -45,6 +45,10 @@ class _ServerSettingsScreenState extends State<ServerSettingsScreen> {
   }
 
   Future<void> savePriorities(String? serverUrl, int newIndex) async {
+    if (_indexCache[serverUrl] != null) {
+      await Provider.of<ElectrumConnection>(context, listen: false)
+          .closeConnection();
+    }
     if (newIndex != _indexCache[serverUrl]) {
       _indexCache[serverUrl] = newIndex;
       _servers[newIndex].setPriority = newIndex;
@@ -85,6 +89,7 @@ class _ServerSettingsScreenState extends State<ServerSettingsScreen> {
           Padding(
             padding: const EdgeInsets.symmetric(horizontal: 10.0),
             child: IconButton(
+              key: const Key('serverSettingsAddServer'),
               onPressed: () async {
                 var result = await Navigator.of(context)
                     .pushNamed(Routes.serverAdd, arguments: _walletName);
@@ -182,11 +187,15 @@ class _ServerSettingsScreenState extends State<ServerSettingsScreen> {
                         margin: const EdgeInsets.symmetric(
                             horizontal: 15, vertical: 4),
                         color: Theme.of(context).errorColor,
-                        child: const Icon(Icons.delete,
-                            color: Colors.white, size: 40),
+                        child: const Icon(
+                          Icons.delete,
+                          color: Colors.white,
+                          size: 40,
+                        ),
                       ),
                       key: Key(_servers[index].address),
                       child: Card(
+                        clipBehavior: Clip.antiAlias,
                         margin: const EdgeInsets.fromLTRB(8, 8, 8, 0),
                         child: ListTile(
                           leading: Column(
@@ -210,15 +219,16 @@ class _ServerSettingsScreenState extends State<ServerSettingsScreen> {
                                       element.connectable == true) ==
                                   null) {
                                 //show snack bar
-                                ScaffoldMessenger.of(context)
-                                    .showSnackBar(SnackBar(
-                                  content: Text(
-                                    AppLocalizations.instance.translate(
-                                        'server_settings_error_no_server_left'),
-                                    textAlign: TextAlign.center,
+                                ScaffoldMessenger.of(context).showSnackBar(
+                                  SnackBar(
+                                    content: Text(
+                                      AppLocalizations.instance.translate(
+                                          'server_settings_error_no_server_left'),
+                                      textAlign: TextAlign.center,
+                                    ),
+                                    duration: const Duration(seconds: 2),
                                   ),
-                                  duration: const Duration(seconds: 2),
-                                ));
+                                );
 
                                 //reset connectable
                                 oldItem.setConnectable = true;
@@ -242,17 +252,21 @@ class _ServerSettingsScreenState extends State<ServerSettingsScreen> {
                                 _servers[index].setPriority = 0;
                               }
                             },
-                            icon: Icon(_servers[index].connectable
-                                ? Icons.offline_bolt
-                                : Icons.offline_bolt_outlined),
+                            icon: Icon(
+                              _servers[index].connectable
+                                  ? Icons.offline_bolt
+                                  : Icons.offline_bolt_outlined,
+                            ),
                           ),
                           tileColor: calculateTileColor(
                               index, _servers[index].connectable),
                           title: Text(_servers[index].address),
                           subtitle: _servers[index].address == _connectedServer
                               ? Center(
-                                  child: Text(AppLocalizations.instance
-                                      .translate('wallet_connected')),
+                                  child: Text(
+                                    AppLocalizations.instance
+                                        .translate('wallet_connected'),
+                                  ),
                                 )
                               : Container(),
                         ),