+++ lib/screens/wallet/wallet_import_scan.dart
@@ -64,8 +64,9 @@ class _WalletImportScanScreenState extends State<WalletImportScanScreen> {
               _coinName, 0, 0, 0, true);
 
           //subscribe to hd master
-          _connectionProvider!.subscribeToScriptHashes(await _activeWallets
-              .getWalletScriptHashes(_coinName, _masterAddr));
+          _connectionProvider!.subscribeToScriptHashes(
+            await _activeWallets.getWalletScriptHashes(_coinName, _masterAddr),
+          );
 
           setState(() {
             _scanStarted = true;